#6 [move] : usecase 디렉토리 변경
