+++ app/src/main/java/com/sopt/now/domain/usecase/sharedprefusecase/SaveCheckLoginUseCase.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.domain.usecase
+package com.sopt.now.domain.usecase.sharedprefusecase
 
 import com.sopt.now.domain.repository.UserInfoRepository
 import javax.inject.Inject
