+++ app/src/main/java/com/sopt/now/domain/usecase/sharedprefusecase/SaveUserInfoUseCase.kt
@@ -1,13 +1,13 @@
-package com.sopt.now.domain.usecase
+package com.sopt.now.domain.usecase.sharedprefusecase
 
 import com.sopt.now.domain.entity.UserEntity
 import com.sopt.now.domain.repository.UserInfoRepository
 import javax.inject.Inject
 
-class GetUserInfoUseCase @Inject constructor(
+class SaveUserInfoUseCase @Inject constructor(
     private val userInfoRepository: UserInfoRepository
 ) {
-    operator fun invoke(): UserEntity {
-        return userInfoRepository.getUserInfo()
+    operator fun invoke(user: UserEntity) {
+        userInfoRepository.saveUserInfo(user)
     }
 }
